refactor: 프로젝트 폴더 구조 개선 및 리팩토링
- `components/` 내부를 `common/`, `books/`, `errorBoundary/`,  `skeleton/`으로 구조화
- `component/books/`에 도서 관련 컴포넌트 이동
- `component/common/`에 공통 컴포넌트 이동
- `skeletons/` → `components/common/Skeletons/`로 이동
