+++ Frontend/src/components/books/BookDetail.tsx
