+++ Frontend/src/components/skeletons/BooksTableSkeleton.tsx
